@@ -1,6 +1,6 @@
 class Api::EventsController < ApplicationController
 
-  before_filter :authenticate
+  before_action :authenticate
 
   respond_to :json
 
@@ -13,7 +13,12 @@ class Api::EventsController < ApplicationController
   end
 
   def create
-    @event = Event.new_with_session(params[:event], session)
+    @event = Event.new(
+      event_params.merge(
+        booked_by_email: session[:email],
+        booked_by_name: session[:name]
+      )
+    )
     unless @event.save
       return render json: @event.errors, status: :unprocessable_entity
     end
@@ -22,7 +27,7 @@ class Api::EventsController < ApplicationController
 
   def update
     @event = get_event
-    unless @event.update_safe_attributes(params[:event])
+    unless @event.update(event_params)
       return render json: @event.errors, status: :unprocessable_entity
     end
     render action: :show
@@ -38,9 +43,17 @@ class Api::EventsController < ApplicationController
     render json: 'Invalid Record', status: :not_found
   end
 
+  rescue_from  ActionController::ParameterMissing do |e|
+    render json: e.message, status: 422
+  end
+
   protected
   def get_event
     Event.owned_by(session[:email]).find(params[:id])
+  end
+
+  def event_params
+    params.require(:event).permit(:description, :date)
   end
 
 end
