@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -9,17 +8,17 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130528005026) do
+ActiveRecord::Schema.define(version: 2013_05_28_005026) do
 
-  create_table "events", :force => true do |t|
-    t.string   "booked_by_email"
-    t.string   "booked_by_name"
-    t.string   "description"
-    t.date     "date"
-    t.datetime "created_at",      :null => false
-    t.datetime "updated_at",      :null => false
+  create_table "events", force: :cascade do |t|
+    t.string "booked_by_email"
+    t.string "booked_by_name"
+    t.string "description"
+    t.date "date"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
 end
